PyTest DQ Framework/requirements.txt
pytest==8.3.5
pytest-html==4.1.1
psycopg2-binary==2.9.10
pandas==2.2.3
pyarrow==21.0.0